Put this analysis in a separate folder such that I can clearly separate it from the original analysis
